refactor: add new calendars colors
