--- frontend/src/utils/common.js
@@ -1,19 +1,0 @@
-export const Color = {
-	ONE: '#E7B431',
-	TWO: '#FF9086',
-	THREE: '#E25757',
-	FOUR: '#E4F2FF',
-	FIVE: '#008EF4',
-	SIX: '#00759A',
-	SEVEN: '#3FBA62',
-	EIGHT: '#225662',
-	NINE: '#7F9498',
-	DEFAULT: '#91DED3',
-};
-
-export const Status = {
-	IDLE: 'idle',
-	LOADING: 'loading',
-	SUCCESS: 'success',
-	ERROR: 'error',
-};
